test: BoardView command 입력 테스트
